feat(headers): parse AbsolutePath further into (path,query,fragment)
This commit parses a RequestUri further in the case of AbsolutePath.
It defines PathQueryFragment aligning to the output of rust-url's
parse_path() function, which is intended for the purpose of parsing
a "standalone" URL path as "typically found in the start line of an
HTTP header".

BREAKING_CHANGE: Change RequestUri::AbsolutePath(String) to RequestUri::AbsolutePath(PathQueryFragment)